pyproject.toml
[project]
name = "mini-cfo-copilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kaleido>=0.2.1",
    "matplotlib>=3.7",
    "numpy>=1.24",
    "pandas>=2.0",
    "pillow>=9.0.0",
    "plotly>=5.15.0",
    "pytest>=7.0",
    "python-dateutil>=2.8.2",
    "reportlab>=3.6.0",
    "streamlit>=1.37",
]
